chore: solve warnings out from variadic macro
Warning: ISO C++11 requires at least one argument for the "..." in a variadic macro
Because of usage of -Wpedantic and not passing any argument to the variadic macro, the compiler will throw a warning.
This commit solves the warning by removing the trailing comma from the variadic macro.

Ticket: SOUR-76